pyproject.toml
[tool.poetry]
name = "calc-shared-gyd-yield"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
asyncio = "^3.4.3"
aiohttp = "^3.9.5"
gql = "^3.5.0"
pandas = "^2.2.2"
numpy = "^2.0.0"
icecream = "^2.1.3"
dune-client = "^1.7.4"
ipykernel = "^6.29.4"
web3 = "^6.19.0"
jsno = "^1.2.3"
jmespath = "^1.0.1"
pyyaml = "^6.0.1"
eth-abi = "^5.1.0"
toolz = "^0.12.1"
tzlocal = "^5.2"


[tool.poetry.group.dev.dependencies]
ptpython = "^3.0.27"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
